requirements.txt
bleak==0.22.2
certifi==2024.8.30
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.8
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
keyring==25.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
nest-asyncio==1.6.0
nh3==0.2.18
pkginfo==1.10.0
Pygments==2.18.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
qasync==0.27.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
setuptools==74.1.2
supertools==1.0.1
twine==5.1.1
urllib3==2.2.3
wheel==0.44.0
winrt-runtime==2.2.0
winrt-Windows.Devices.Bluetooth==2.2.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.2.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.2.0
winrt-Windows.Devices.Enumeration==2.2.0
winrt-Windows.Foundation==2.2.0
winrt-Windows.Foundation.Collections==2.2.0
winrt-Windows.Storage.Streams==2.2.0
zipp==3.20.2